Responsive Design (Pull Request #2)
* Added gitignore

* Removed redundant code

This should not be in the github files. Anyone could view and access the db. Move this to an environment variable

* Added holders for new pages

* Moved NavBar out of AppContext

Moving it out allowed it to be placed above the content (not floating). All pages will need to be modified to cater for this change

* Updated Homepage for mobile and readability

Added glass (from daisyui) to allow content to be placed on images nicely

* Added Navbar to all pages

* Responsive Stakeholder Design

* Made Case Studies Responsive

* Added a shadow to distinguish between menu and content

* Updated Project-Stories

* Update Team page

* Updated Contact page

* update ExploreProjects layout

* Update User Login and Register

* Handle if text is undefined

* Set input fields to password

* Updated NewApplication Page

* Updated ProjectView for mobile